requirements.txt
litellm>=1.0.0
python-dotenv>=0.21.0
requests>=2.28.0
google-generativeai>=0.3.0
markdown>=3.4.0
jinja2>=3.1.0